Move get published into next milestone component
