+++ src/screens/Overview/components/NextMilestone/components/GetPublished/components/GetPublishedSteps/index.js
