+++ src/screens/Overview/components/NextMilestone/components/GetPublished/components/GetPublishedSteps/utils/isStepComplete/index.test.js
