+++ src/screens/Overview/components/NextMilestone/index.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+// TODO: Add milestones (no published lessons -> get published, < 12 lessons -> unlock self approval)
+
+export default ({lessonPage}) => (
+  <div></div>
+)
