+++ src/screens/Overview/index.js
@@ -1,7 +1,8 @@
 import React from 'react'
 import {overviewTitleText} from 'utils/text'
 import Split from 'components/Split'
-import Intro from './components/Intro'
+import Hello from './components/Hello'
+import NextMilestone from './components/NextMilestone'
 import Lessons from './components/Lessons'
 import Stats from './components/Stats'
 import Help from './components/Help'
@@ -12,14 +13,17 @@ export default ({
 }) => (
   <Split
     intro={
-      <Intro instructor={instructor} />
+      <Hello instructor={instructor} />
     }
     title={overviewTitleText}
     main={
-      <Lessons
-        instructor={instructor}
-        lessonPage={lessonPage}
-      />
+      <div>
+        <NextMilestone lessonPage={lessonPage} />
+        <Lessons
+          instructor={instructor}
+          lessonPage={lessonPage}
+        />
+      </div>
     }
     aside={
       <div>
